chore(`ci`): add `ci-success` step so we can make this a condition for merging (#111)
<!--
Thank you for your Pull Request. Please provide a description above and
review
the requirements below.

Bug fixes and new features should include tests.

Contributors guide:
https://github.com/foundry-rs/foundry-fork-db/blob/main/CONTRIBUTING.md

The contributors guide includes instructions for running rustfmt and
building the
documentation.
-->

<!-- ** Please select "Allow edits from maintainers" in the PR Options
** -->

## Motivation

<!--
Explain the context and why you're making that change. What is the
problem
you're trying to solve? In some cases there is not a problem and this
can be
thought of as being the motivation for your change.
-->

Adds `ci-success` similar to how we have in Foundry already to allow
selecting it as a condition for merging

## Solution

<!--
Summarize the solution and provide any necessary context needed to
understand
the code change.
-->

A PR will follow to further harden the workflow